Fix ARM cross-compilation and build system
- CMakeLists.txt: add ai_camera to install targets, set output dir to
  build root, fix glob for fftw/spdlog/bedmesh .so variants
- common/utilities.h: add #include <cstdint> for ARM GCC 6.4.1
- dsp_helper.h, serialhdl.cpp: remove conflicting asm-generic/termbits.h
  kernel header include (conflicts with glibc termios on ARM)
- lib/curl/bin/arm/libcurl.so{,.4}: replace linker scripts with symlinks
- lib/openssl/bin/arm/libssl.so, libcrypto.so: replace linker scripts
  with symlinks (GCC 6.4.1 rejects libtool-style linker scripts)
- Warning fixes across extras/, kinematics/, chelper/ for clean builds
  on both x86_64 and ARM toolchains

+++ elegoo/chelper/stepcompress.c
@@ -210,7 +210,7 @@ check_line(struct stepcompress *sc, struct step_move move)
     if (!move.count || (!move.interval && !move.add && move.count > 1)
         || move.interval >= 0x80000000) {
         errorf("stepcompress o=%d i=%u c=%d a=%d last_clock=%llu queue_pos=%u, : Invalid sequence"
-               , sc->oid, move.interval, move.count, move.add, sc->last_step_clock, *sc->queue_pos);
+               , sc->oid, move.interval, move.count, move.add, (unsigned long long)sc->last_step_clock, *sc->queue_pos);
         return ERROR_RET;
     }
     uint32_t interval = move.interval, p = 0;
